Deprecate methods from TableTools. Add additional methods to CsvHelpers.
